CCAP-591, Replacing BasicApplicationRouterService with ApplicationRouterService and using database driven resource org, zipcode

+++ src/main/java/org/ilgcc/app/data/ResourceOrganizationRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ResourceOrganizationRepository extends JpaRepository<ResourceOrganization, Long> {
+public interface ResourceOrganizationRepository extends JpaRepository<ResourceOrganization, BigInteger> {
 
     List<ResourceOrganization> findByCaseloadCode(String caseloadCode);
